ab_riscv_primitives::instructions::rv32::zce::zcmp

Trait ZcmpRegister

pub unsafe trait ZcmpRegister
where
    Self: Register,
{
    const RVE: bool;
}

General purpose register with additional constraints for Zcmp extension

Safety

Register::from_bits() must return Some() for:

1, 8, 9 and 18..=27 if Self::RVE = false

1, 8 and 9 if Self::RVE = true

Required Associated Constants

const RVE: bool

Whether this is RVE variant with the number of general purpose registers reduced to 16

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".